% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-params.R
\name{.is_scalar_finite}
\alias{.is_scalar_finite}
\title{Test whether \code{x} is a finite length-1 numeric}
\usage{
.is_scalar_finite(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} / \code{FALSE}.
}
\description{
Test whether \code{x} is a finite length-1 numeric
}
\keyword{internal}
