% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-engine.R
\name{.S_lnorm}
\alias{.S_lnorm}
\title{Lognormal survivor function S(t)}
\usage{
.S_lnorm(t, mu, sigma)
}
\arguments{
\item{t}{Numeric vector of times.}

\item{mu}{Numeric scalar; lognormal location parameter.}

\item{sigma}{Numeric scalar; lognormal scale parameter.}
}
\value{
Numeric vector the same length as \code{t}.
}
\description{
Lognormal survivor function S(t)
}
\keyword{internal}
