% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-engine.R
\name{.S_weibull_PH}
\alias{.S_weibull_PH}
\title{Weibull (proportional-hazards parameterisation) survivor function S(t)}
\usage{
.S_weibull_PH(t, ln_lambda, gamma)
}
\arguments{
\item{t}{Numeric vector of times.}

\item{ln_lambda}{Numeric scalar; log of the rate parameter.}

\item{gamma}{Numeric scalar; shape parameter.}
}
\value{
Numeric vector the same length as \code{t}.
}
\description{
Weibull (proportional-hazards parameterisation) survivor function S(t)
}
\keyword{internal}
