% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moon-params.R
\name{is_moon_param}
\alias{is_moon_param}
\title{Test whether an object is a MOON parameter spec}
\usage{
is_moon_param(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{"moon_param"}, otherwise \code{FALSE}.
}
\description{
Test whether an object is a MOON parameter spec
}
\keyword{internal}
