% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moon-package.R
\docType{package}
\name{moon-package}
\alias{moon}
\alias{moon-package}
\title{moon: Modeling Obesity in Norway (MOON) Markov Cohort Model}
\description{
Implements the MOON Markov cohort model for simulating obesity prevalence, incremental health-care costs, and years of life lost in a Norwegian birth cohort over the life course. Originally described in Bjørnelv et al. (2021) \doi{10.1177/0272989X20971589}; this package is an R port and refactor of the Excel/VBA implementation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DaCHE-SDU/moon}
  \item \url{https://dache-sdu.github.io/moon}
  \item Report bugs at \url{https://github.com/DaCHE-SDU/moon/issues}
}

}
\author{
\strong{Maintainer}: Nicolai Damslund \email{nsimonsen@health.sdu.dk} (\href{https://orcid.org/0000-0002-4649-3109}{ORCID})

Authors:
\itemize{
  \item Nicolai Damslund \email{nsimonsen@health.sdu.dk} (\href{https://orcid.org/0000-0002-4649-3109}{ORCID})
  \item Gudrun M W Bjørnelv (\href{https://orcid.org/0000-0003-4997-5426}{ORCID}) (Original MOON model (Bjørnelv et al. 2021, doi:10.1177/0272989X20971589))
  \item Yufei Jiang \email{yjiang@health.sdu.dk} (\href{https://orcid.org/0009-0004-5389-7971}{ORCID}) (Ported the MOON model from the original Excel/VBA implementation to R)
  \item Magnus Skov Larsen \email{maskla@health.sdu.dk} (\href{https://orcid.org/0009-0002-7481-734X}{ORCID}) (Ported the MOON model from the original Excel/VBA implementation to R)
  \item Christina Hansen Edwards \email{christina.h.edwards@ntnu.no} (\href{https://orcid.org/0000-0003-2329-1739}{ORCID})
  \item Vidar Halsteinli (\href{https://orcid.org/0000-0002-1052-1737}{ORCID}) (Original MOON model (Bjørnelv et al. 2021, doi:10.1177/0272989X20971589))
  \item Rønnaug A Ødegaard (Original MOON model (Bjørnelv et al. 2021, doi:10.1177/0272989X20971589))
  \item Line Planck Kongstad \email{lpkongstad@health.sdu.dk} (\href{https://orcid.org/0000-0002-9625-5870}{ORCID})
}

}
\keyword{internal}
